+++ packages/esm-patient-test-results-app/src/grouped-timeline/index.ts
@@ -1,0 +1,5 @@
+import GroupedTimeline from './grouped-timeline';
+import styles from './grouped-timeline.styles.scss';
+
+export default GroupedTimeline;
+export { GroupedTimeline, styles };
