+++ packages/esm-patient-test-results-app/src/results-viewer/index.ts
@@ -1,0 +1,4 @@
+import ResultsViewer from './results-viewer';
+
+export default ResultsViewer;
+export { ResultsViewer };
