+++ packages/esm-patient-test-results-app/src/timeline/index.ts
@@ -1,3 +1,3 @@
-import Timeline from './Timeline';
+import Timeline from './timeline.component';
 
 export default Timeline;
